TASK: Rename `CreationDialogPostprocessor` to `CreationDialogNodeTypePostprocessor`
... and colocate code with `PromotedElementsCreationHandler`
